+++ scripts/exporters/abstract-exporter.mjs
@@ -172,6 +172,12 @@ export class AbstractExporter {
       Hooks.callAll('BTFG.afterDocumentProcessed', indexDocument, documentData, this.options, this.#manager);
       Hooks.callAll(`BTFG.after${this.pack.metadata.type}DocumentProcessed`, indexDocument, documentData, this.options, this.#manager);
 
+      if (['Actor', 'Item'].includes(this.pack.metadata.type)) {
+        const subType = indexDocument.type.charAt(0).toUpperCase() + indexDocument.type.slice(1);
+
+        Hooks.callAll(`BTFG.after${this.pack.metadata.type}${subType}DocumentProcessed`, indexDocument, documentData, this.options, this.#manager);
+      }
+
       this.dataset.entries[key] = foundry.utils.mergeObject(
         documentData,
         this.existingContent[key] ?? {},